+++ ChangeLog.md
@@ -4,7 +4,7 @@ Version 4.94 - Atmo autopilot overhaul
 - Better yawing to target to not stall when at low speeds and proceeding to waypoint
 - Will now roll to turn when at high speeds (>100m/s) and proceeding to waypoint, respecting stall limits
 - Improved waypoint accuracy with better yaw/roll convergence
-- Brake Landings will use BrakeLandingRate when initiated with G, or calculated brake landing rate if used with autopilot at a saved landing location (height known).  Calculated is much faster.
+- New User Variable: CalculateBrakeLandingSpeed = false --export: (Default: false) Whether BrakeLanding speed at non-waypoints should be calculated or use the BrakeLandingRate user value
 
 Version 4.935
 - Extremely dangerous BrakeLanding changes.  Brakelanding is now, again, faster - but attempts to put you at effectively 0m above the ground.  Please let me know if this breaks your ship so I can adjust it, but it works fine on all of mine.  May cause issues if things are under your ship when landing and aren't detected by your hovers/vboosters